+++ tests/testthat/test-utility.R
@@ -56,6 +56,50 @@ test_that("sf6d_utility works with Australian value set", {
   expect_equal(utility, c(0.193, 0.337))
 })
 
+test_that("sf6d_utility works with Hong Kong value set", {
+  responses <- list(Q1 = 1:2, Q2 = 1:2, Q3 = 1:2, Q4 = 1:2, Q5 = 1:2, Q6 = 1:2, Q7 = 1:2, Q8 = 1:2,
+                    Q9 = 1:2, Q10 = 1:2, Q11 = 1:2, Q12 = 1:2, Q13 = 1:2, Q14 = 1:2, Q15 = 1:2,
+                    Q16 = 1:2, Q17 = 1:2, Q18 = 1:2, Q19 = 1:2, Q20 = 1:2, Q21 = 1:2, Q22 = 1:2,
+                    Q23 = 1:2, Q24 = 1:2, Q25 = 1:2, Q26 = 1:2, Q27 = 1:2, Q28 = 1:2, Q29 = 1:2,
+                    Q30 = 1:2, Q31 = 1:2, Q32 = 1:2, Q33 = 1:2, Q34 = 1:2, Q35 = 1:2, Q36 = 1:2)
+  utility <- rlang::exec(sf6d_utility, !!!responses, questionnaire = "SF-36", values = "hk")
+
+  expect_equal(utility, c(0.472, 0.464))
+})
+
+test_that("sf6d_utility works with ERUM value set", {
+  responses <- list(Q1 = 1:2, Q2 = 1:2, Q3 = 1:2, Q4 = 1:2, Q5 = 1:2, Q6 = 1:2, Q7 = 1:2, Q8 = 1:2,
+                    Q9 = 1:2, Q10 = 1:2, Q11 = 1:2, Q12 = 1:2, Q13 = 1:2, Q14 = 1:2, Q15 = 1:2,
+                    Q16 = 1:2, Q17 = 1:2, Q18 = 1:2, Q19 = 1:2, Q20 = 1:2, Q21 = 1:2, Q22 = 1:2,
+                    Q23 = 1:2, Q24 = 1:2, Q25 = 1:2, Q26 = 1:2, Q27 = 1:2, Q28 = 1:2, Q29 = 1:2,
+                    Q30 = 1:2, Q31 = 1:2, Q32 = 1:2, Q33 = 1:2, Q34 = 1:2, Q35 = 1:2, Q36 = 1:2)
+  utility <- rlang::exec(sf6d_utility, !!!responses, questionnaire = "SF-36", values = "erum")
+
+  expect_equal(utility, c(0.338, 0.543))
+})
+
+test_that("sf6d_utility works with Spanish value set", {
+  responses <- list(Q1 = 1:2, Q2 = 1:2, Q3 = 1:2, Q4 = 1:2, Q5 = 1:2, Q6 = 1:2, Q7 = 1:2, Q8 = 1:2,
+                    Q9 = 1:2, Q10 = 1:2, Q11 = 1:2, Q12 = 1:2, Q13 = 1:2, Q14 = 1:2, Q15 = 1:2,
+                    Q16 = 1:2, Q17 = 1:2, Q18 = 1:2, Q19 = 1:2, Q20 = 1:2, Q21 = 1:2, Q22 = 1:2,
+                    Q23 = 1:2, Q24 = 1:2, Q25 = 1:2, Q26 = 1:2, Q27 = 1:2, Q28 = 1:2, Q29 = 1:2,
+                    Q30 = 1:2, Q31 = 1:2, Q32 = 1:2, Q33 = 1:2, Q34 = 1:2, Q35 = 1:2, Q36 = 1:2)
+  utility <- rlang::exec(sf6d_utility, !!!responses, questionnaire = "SF-36", values = "spain")
+
+  expect_equal(utility, c(0.160, 0.498))
+})
+
+test_that("sf6d_utility works with Portuguese value set", {
+  responses <- list(Q1 = 1:2, Q2 = 1:2, Q3 = 1:2, Q4 = 1:2, Q5 = 1:2, Q6 = 1:2, Q7 = 1:2, Q8 = 1:2,
+                    Q9 = 1:2, Q10 = 1:2, Q11 = 1:2, Q12 = 1:2, Q13 = 1:2, Q14 = 1:2, Q15 = 1:2,
+                    Q16 = 1:2, Q17 = 1:2, Q18 = 1:2, Q19 = 1:2, Q20 = 1:2, Q21 = 1:2, Q22 = 1:2,
+                    Q23 = 1:2, Q24 = 1:2, Q25 = 1:2, Q26 = 1:2, Q27 = 1:2, Q28 = 1:2, Q29 = 1:2,
+                    Q30 = 1:2, Q31 = 1:2, Q32 = 1:2, Q33 = 1:2, Q34 = 1:2, Q35 = 1:2, Q36 = 1:2)
+  utility <- rlang::exec(sf6d_utility, !!!responses, questionnaire = "SF-36", values = "portugal")
+
+  expect_equal(utility, c(0.557, 0.793))
+})
+
 test_that("sf6d_utility generates appropriate errors", {
   expect_error(sf6d_utility(questionnaire = "X"), class = "SF6Dvalues_error_invalid_version")
   expect_error(sf6d_utility(values = "nz"), class = "SF6Dvalues_error_not_implemented")
@@ -63,6 +107,26 @@ test_that("sf6d_utility generates appropriate errors", {
   expect_error(
     sf6d_utility(Q1 = 1, Q2 = 1, Q3 = 1, Q4 = 1, Q5 = 1, Q6 = 1, Q7 = 1, Q8 = 1, Q9 = 1, Q10 = 1,
                  Q11 = 1, Q12 = 1, values = "oz"),
+    class = "SF6Dvalues_error_incompatible_values"
+  )
+  expect_error(
+    sf6d_utility(Q1 = 1, Q2 = 1, Q3 = 1, Q4 = 1, Q5 = 1, Q6 = 1, Q7 = 1, Q8 = 1, Q9 = 1, Q10 = 1,
+                 Q11 = 1, Q12 = 1, values = "hk"),
+    class = "SF6Dvalues_error_incompatible_values"
+  )
+  expect_error(
+    sf6d_utility(Q1 = 1, Q2 = 1, Q3 = 1, Q4 = 1, Q5 = 1, Q6 = 1, Q7 = 1, Q8 = 1, Q9 = 1, Q10 = 1,
+                 Q11 = 1, Q12 = 1, values = "erum"),
+    class = "SF6Dvalues_error_incompatible_values"
+  )
+  expect_error(
+    sf6d_utility(Q1 = 1, Q2 = 1, Q3 = 1, Q4 = 1, Q5 = 1, Q6 = 1, Q7 = 1, Q8 = 1, Q9 = 1, Q10 = 1,
+                 Q11 = 1, Q12 = 1, values = "spain"),
+    class = "SF6Dvalues_error_incompatible_values"
+  )
+  expect_error(
+    sf6d_utility(Q1 = 1, Q2 = 1, Q3 = 1, Q4 = 1, Q5 = 1, Q6 = 1, Q7 = 1, Q8 = 1, Q9 = 1, Q10 = 1,
+                 Q11 = 1, Q12 = 1, values = "portugal"),
     class = "SF6Dvalues_error_incompatible_values"
   )
   expect_warning(