+++ NEWS.md
@@ -1,8 +1,11 @@
 # SF6Dvalues 0.4.0.9000
 
+* New `hk()`, `erum()`, `spain()`, and `portugal()` SF-6D(SF-36) value sets.
+
 # SF6Dvalues 0.4.0
 
-* `SF6D.as_SF6D.SF6Dvalues_SF36()` now correctly calculates SF-6D(SF-36) profile.
+* `SF6D.as_SF6D.SF6Dvalues_SF36()` now correctly calculates SF-6D(SF-36)
+  profile.
 
 * New `SF36_scores()` calculates SF-36 domain scale scores.